refactored flow
