+++ packages/kokoas-client/src/pages/projContractV2/parts/preview/startcontract/customNodes/nodeTypes.ts
@@ -1,0 +1,7 @@
+import { CircleNode } from './CircleNode';
+import { RoleNode } from './RoleNode';
+
+export const nodeTypes =  ({ 
+  roleNode: RoleNode, 
+  circleNode: CircleNode,
+});
